% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/re_functions2.r
\name{coords2country}
\alias{coords2country}
\title{coords2country finds country for a given coordinate}
\usage{
coords2country(points)
}
\arguments{
\item{points}{coordinates to get coordinate from}
}
\description{
coords2country finds country for a given coordinate
}

