% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/re_functions2.r
\name{find.origin}
\alias{find.origin}
\title{finds the origin for a set of Populations}
\usage{
find.origin(coords, psi, xlen = 50, ylen = 50, doPlot = F, doPdf = F,
  ...)
}
\description{
finds the origin for a set of Populations
}

