% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/re_functions2.r
\name{get.all.psi}
\alias{get.all.psi}
\title{calculates the psi matrix for a pop object}
\usage{
get.all.psi(pop, n = 2, subset = NULL, resampling = "hyper")
}
\arguments{
\item{pop}{population data object from make.pop}

\item{n}{the sample size which we downsample to}

\item{resampling}{mode of resampling. Currently, only
hyper is supported}
}
\value{
A n x n matrix of psi values
}
\description{
calculates the psi matrix for a pop object
}
\examples{
\dontrun{
f = system.file("examples/example.data.bed", package="rangeExpansion")
c = system.file("examples/example.coordinates.csv", package="rangeExpansion")
coord.data <- load.coord.file(c, sep=",")
snp.data <- load.plink.file(f, sep=",")
raw.data <- check.missing(snp.data, coord.data)
raw.data <- set.outgroups(raw.data)
pop <- make.pop(raw.data)
psi <- get.all.psi(pop)
}
}

