% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/re_functions2.r
\name{get.psi}
\alias{get.psi}
\title{the psi statistic calculation
the actual calculation of the psi statistic for
a single pair of populations. The function requires
4 vectors, all of length equal to the number of snps
to be analyzed.}
\usage{
get.psi(ni, nj, fi, fj, n = 2, resampling = "hyper")
}
\arguments{
\item{ni}{the number of sampled haplotypes in population i}

\item{nj}{the number of sampled haplotypes in population j}

\item{fi}{the number of derived alleles in population i}

\item{fj}{the number of derived alleles in population j}

\item{n}{the number of samples to downsample to}
}
\value{
psi a matrix of pairwise psi values
}
\description{
the psi statistic calculation
the actual calculation of the psi statistic for
a single pair of populations. The function requires
4 vectors, all of length equal to the number of snps
to be analyzed.
}
\examples{


example.data <- system.file("examples/example_snp.snapp",
                            package="rangeExpansion")
example.coords <- system.file("examples/example_coordinates.csv",
                            package="rangeExpansion")

raw.data <- load.data.snapp(example.data,
                            example.coords,
                            sep=',', ploidy=ploidy)                                
region <- list(NULL, c("REGION_1", "REGION_2"))

pop <- make.pop(raw.data, ploidy)
psi <- get.all.psi(pop)

res <- run.regions(region=region, pop=pop, psi=psi, xlen=10,ylen=20)

summary(res)
plot(res)
}

