% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/re_functions2.r
\name{get.sample.bbox}
\alias{get.sample.bbox}
\title{gets a bounding box around the populations in samples}
\usage{
get.sample.bbox(samples)
}
\description{
gets a bounding box around the populations in samples
}

