% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/re_functions2.r
\name{hets}
\alias{hets}
\title{Heterozygosity function}
\usage{
hets(x)
}
\arguments{
\item{x}{the allele frequency to calculate heterozygosity for}
}
\value{
Heterozygosity = x(1-x)
}
\description{
Heterozygosity function
}
\examples{
freqs <- 0:10/10
h <- hets(freqs)
}

