% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/re_functions2.r
\name{load.data}
\alias{load.data}
\title{loads a data set from a plink and coords file}
\usage{
load.data(plink.file, coords.file, n.snp = NULL, ploidy = 2, ...)
}
\arguments{
\item{plink.file}{the name of a plink bed file, passed to read.plink}

\item{coords.file}{a file containing coordinates and other info
about the sample}

\item{n.snp}{the maximum number of snp to be read. if NULL, all
snps are read}

\item{ploidy}{the ploidy of the organism, 1 for haploids, 2 for diploid}

\item{...}{further arguments passed to load.coord.file}
}
\value{
an object of type origin.data with entries `genotypes`,
     containing the genetic data and entry `coord` containing location
     data
}
\description{
This is the main function to load a data set. It requires a
plink file in bed format and a coords file
}

