% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/re_functions2.r
\name{load.data.snapp}
\alias{load.data.snapp}
\title{loads a data set from a snapp and coords file}
\usage{
load.data.snapp(snapp.file, coords.file, n.snp = -1, ploidy = 2, ...)
}
\arguments{
\item{snapp.file}{a file in snapp data format}

\item{coords.file}{a file containing coordinates and other info
about the sample}

\item{n.snp}{the maximum number of snp to be read. if -1, all
snps are read}

\item{ploidy}{the ploidy of the organism, 1 for haploids, 2 for diploid}

\item{...}{further arguments passed to load.coord.file}
}
\value{
an object of type origin.data with entries `genotypes`,
     containing the genetic data and entry `coord` containing location
     data
}
\description{
This is a function that reads the old data format. Presented for backwards
compatibility
}
\examples{


example.data <- system.file("examples/example_snp.snapp",
                            package="rangeExpansion")
example.coords <- system.file("examples/example_coordinates.csv",
                            package="rangeExpansion")

raw.data <- load.data.snapp("examples/example_snp.snapp", 
                            "examples/example_coordinates.csv", 
                            sep=',', ploidy=ploidy)                                
region <- list(NULL, c("REGION_1", "REGION_2"))

pop <- make.pop(raw.data, ploidy)
psi <- get.all.psi(pop)

res <- run.regions(region=region, pop=pop, psi=psi, xlen=10,ylen=20)

summary(res)
plot(res)
}

