% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/re_functions2.r
\name{load.plink.file}
\alias{load.plink.file}
\title{Wrapper around snpStats::read.plink to read data in bed/bim/fam format}
\usage{
load.plink.file(..., n.snp = NULL)
}
\description{
Data can be converted to bed/bim/fam format e.g.
using plink 1.9 or bcftools. A useful command line
to use is
\code{plink --vcf my.vcf --set-missing-var-ids "@.#" --make-bed --out new --allow-extra-chr}
}
\examples{
\dontrun{
f = system.file("examples/example.data.bed", package="rangeExpansion")
load.plink.file(f, sep=",")
}
}

