% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/re_functions2.r
\name{load.snapp.file}
\alias{load.snapp.file}
\title{Reads data from SNAPP file}
\usage{
load.snapp.file(snp.file = "../example_data/example_snp.snapp", n.snp = -1)
}
\description{
Deprecated, use plink input format instead
}
\examples{


example.data <- system.file("examples/example_snp.snapp",
                            package="rangeExpansion")
example.coords <- system.file("examples/example_coordinates.csv",
                            package="rangeExpansion")

raw.data <- load.data.snapp("examples/example_snp.snapp", 
                            "examples/example_coordinates.csv", 
                            sep=',', ploidy=ploidy)                                
region <- list(NULL, c("REGION_1", "REGION_2"))

pop <- make.pop(raw.data, ploidy)
psi <- get.all.psi(pop)

res <- run.regions(region=region, pop=pop, psi=psi, xlen=10,ylen=20)

summary(res)
plot(res)
}

