% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/re_functions2.r
\name{make.pop}
\alias{make.pop}
\title{Sets up the popultion structure data sets}
\usage{
make.pop(raw.data, ploidy = 2)
}
\arguments{
\item{raw.data}{Output from set.outgroups or check.data}

\item{ploidy}{the ploidy of organisms}
}
\value{
A list with entries
- data : an n x p matrix with derived allele counts
- coords : an p x 3 matrix of coords
- n : number of populations
}
\description{
Groups up raw data into a data structure that
}
\examples{
\dontrun{
f = system.file("examples/example.data.bed", package="rangeExpansion")
c = system.file("examples/example.coordinates.csv", package="rangeExpansion")
coord.data <- load.coord.file(c, sep=",")
snp.data <- load.plink.file(f, sep=",")
raw.data <- check.missing(snp.data, coord.data)
raw.data <- set.outgroups(raw.data)
pop <- make.pop(raw.data)
}
}

