% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/re_functions2.r
\name{model.1d}
\alias{model.1d}
\title{calculates distance from a given point}
\usage{
model.1d(xy, data, pct = 0.01, f.dist = "haversine")
}
\arguments{
\item{xy}{coordinate of point to evalute function at}

\item{data}{a 5 col data frame with columns xi, yi, xj, yj, psi, with
coords from the two sample location and their psi statistic.
best generated unsg prep.tdoa.data}

\item{f.dist}{the distance function to use. the default 'haversine'
uses the haversine distance. Alternatively, 'euclidean' uses
Euclidean distance}
}
\value{
an object of type lm describing fit
}
\description{
calculates distance from a given point
}

