% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/re_functions2.r
\name{plot.origin.result.list}
\alias{plot.origin.result.list}
\title{plots a set of origin-inference results}
\usage{
plot.origin.result.list(res, ...)
}
\arguments{
\item{res}{an object of type origin.result.list}

\item{...}{further objects passed. to plot.origin.result}
}
\description{
simply loops over all different regions and calls the plot function
for all of them
}

