% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/re_functions2.r
\name{prep.tdoa.data}
\alias{prep.tdoa.data}
\title{preps data for tdoa analysis}
\usage{
prep.tdoa.data(coords, psi)
}
\arguments{
\item{pop.coords}{coordination data with latitude and longitude}

\item{all.psi}{matrix with psi values}

\item{region}{region or set of regions to run analysis on}

\item{countries}{countries to restrict analysis to}

\item{xlen}{number of x points}

\item{ylen}{number of y points}
}
\description{
makes a data structure of format xi, yi, xj, yj, psi
}

