% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/re_functions2.r
\name{run.analysis}
\alias{run.analysis}
\title{Runs the entire analysis.}
\usage{
run.analysis(snp.file, coord.file, sep = ",", ploidy = 2,
  pop.mode = "coord", regions = NULL, n.points = 20, n.snp = NULL)
}
\arguments{
\item{snp.file}{the bed file with genetic data}

\item{coord.file}{file with coordinate info}

\item{pop.mode}{How populations are created. Default is that individuals
with same location are merged}

\item{regions}{Independent regions for which analysis is supposed to be ran}

\item{n.points}{How many points should be generated for plotting}

\item{n.snp}{Maximum number of snp to be used}
}
\description{
See individual functions for details
}

