% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/re_functions2.r
\name{run.regions}
\alias{run.regions}
\title{runs TDOA analysis for multiple regions}
\usage{
run.regions(region, ...)
}
\arguments{
\item{...}{further arguments passed to plot}

\item{a}{list of vectors, where each vector contains elements
that are compared to pop@coords@region
to determine whether they should be included
if null, all are run}

\item{pop}{population data object from make.pop}

\item{psi}{a matrix of psi values}

\item{output.file.name}{name of output file}

\item{xlen}{number of points to evaluate function in x direction}

\item{ylen}{number of points to evaluate function in y direction}
}
\description{
runs the analysis for all individuals in a region. See
run.single.region for further details on argumetns
}

