% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/re_functions2.r
\name{run.single.region}
\alias{run.single.region}
\title{runs analysis for a single region}
\usage{
run.single.region(region, pop, psi, output.file.name = "default",
  xlen = 100, ylen = 100, bbox = NULL, ...)
}
\arguments{
\item{region}{a vector of entries that are compared to pop@coords@region
to determine whether they should be included}

\item{pop}{population data object from make.pop}

\item{psi}{a matrix of psi values}

\item{xlen}{number of points to evaluate function in x direction}

\item{ylen}{number of points to evaluate function in y direction}

\item{bbox}{type of the bounding box to limit region}

\item{...}{further arguments passed to plot}

\item{loc.file.id}{output stuff}
}
\description{
runs the analysis for all individuals in a region
}

