% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/re_functions2.r
\name{single.origin}
\alias{single.origin}
\title{Finds origin for a region}
\usage{
single.origin(tdoa.data, bbox, pop.coords, pct = 0.01, xlen = 100,
  ylen = 100, exclude.ocean = T, exclude.land = F, ...)
}
\arguments{
\item{tdoa.data}{Object of type tdoa.data}

\item{bbox}{bounding box describing the location where inference should
be done in}

\item{pop.coords}{coordinations}

\item{pct}{threshold parameter in model.1d}

\item{xlen,}{ylen parameters describing the number of points to use}

\item{exclude.land}{boolean, whether points on land should be
excluded}

\item{exclude.ocoean}{boolean, whether points not on land should be
excluded}
}
\description{
Finds origin for a region
}

