% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/re_functions2.r
\name{summary.origin.result.list}
\alias{summary.origin.result.list}
\title{summarizes origin.result.list object
given output from origin inference, returns a table with statistics}
\usage{
summary.origin.result.list(res)
}
\arguments{
\item{res}{an object of type origin.results to be summarized}
}
\value{
a data frame with some summary statistics and the
   most likely origin
}
\description{
summarizes origin.result.list object
given output from origin inference, returns a table with statistics
}

