% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/re_functions2.r
\name{summary.origin.results}
\alias{summary.origin.results}
\title{given output from single.origin, finds the origin}
\usage{
\method{summary}{origin.results}(res)
}
\arguments{
\item{res}{an object of type origin.results to be summarized}
}
\value{
a data frame with some summary statistics and the
   most likely origin
}
\description{
given output from single.origin, finds the origin
}

