% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/re_functions2.r
\name{test.ibd}
\alias{test.ibd}
\title{tests for isolation by distance by doing permutations on data}
\usage{
test.ibd(data, all.psi, pops, n.permutations = 1000)
}
\description{
tests for isolation by distance by doing permutations on data
}

