% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/re_functions2.r
\name{check.missing}
\alias{check.missing}
\title{Cleans genotype and coordinate files for consistency}
\usage{
check.missing(snp.data, coord.data)
}
\arguments{
\item{snp.data}{output from load.plink.file}

\item{coord.data}{output from load.coord.file}
}
\value{
An object with entries 'genotypes' and 'coords' containing the
   input data sets
}
\description{

}
\details{
This is a generic function: methods can be defined for it directly
or via the \code{\link{Summary}} group generic. For this to work properly,
the arguments \code{...} should be unnamed, and dispatch is on the
first argument.
}
\examples{


example.data <- system.file("examples/example_snp.snapp",
                            package="rangeExpansion")
example.coords <- system.file("examples/example_coordinates.csv",
                            package="rangeExpansion")

raw.data <- load.data.snapp(example.data,
                            example.coords,
                            sep=',', ploidy=ploidy)                                
region <- list(NULL, c("REGION_1", "REGION_2"))

pop <- make.pop(raw.data, ploidy)
psi <- get.all.psi(pop)

res <- run.regions(region=region, pop=pop, psi=psi, xlen=10,ylen=20)

summary(res)
plot(res)
}

