% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/re_functions2.r
\name{load.coord.file}
\alias{load.coord.file}
\title{Loads a coordinate data set}
\usage{
load.coord.file(file, ...)
}
\arguments{
\item{...}{Additional arguments passed to read.table}

\item{file:}{the name of the file which the data are to be read from.}
}
\value{
A data.frame with columns sample, longitude, latitude, region,
outgroup and country
}
\description{
Loads a file that specifies location data and other sample specific information
A
}
\examples{
f = system.file("examples/example.coordinates.csv", package="rangeExpansion")
load.coord.file(f, sep=",")
}

