% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/re_functions2.r
\name{load.coord.file}
\alias{load.coord.file}
\title{Loads a coordinate data set}
\usage{
load.coord.file(file, ...)
}
\arguments{
\item{...}{Additional arguments passed to read.table}

\item{file:}{the name of the file which the data are to be read from.}
}
\value{
A data.frame with columns sample, longitude, latitude, region,
outgroup and country
}
\description{
Loads a file that specifies location data and other sample specific information
A
}
\examples{


example.data <- system.file("examples/example_snp.snapp",
                            package="rangeExpansion")
example.coords <- system.file("examples/example_coordinates.csv",
                            package="rangeExpansion")

region <- list(NULL, c("REGION_1", "REGION_2"))
ploidy <- 2
raw.data <- load.data.snapp(example.data,
                            example.coords,
                            sep=',', ploidy=ploidy)                                

pop <- make.pop(raw.data, ploidy)
psi <- get.all.psi(pop)

res <- run.regions(region=region, pop=pop, psi=psi, xlen=10,ylen=20)

summary(res)
plot(res)
}

