% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/re_functions2.r
\name{make.pop}
\alias{make.pop}
\title{Sets up the popultion structure data sets}
\usage{
make.pop(raw.data, ploidy = 2)
}
\arguments{
\item{raw.data}{Output from set.outgroups or check.data}

\item{ploidy}{the ploidy of organisms}
}
\value{
A list with entries
- data : an n x p matrix with derived allele counts
- coords : an p x 3 matrix of coords
- n : number of populations
}
\description{
Groups up raw data into a data structure that
}
\examples{


example.data <- system.file("examples/example_snp.snapp",
                            package="rangeExpansion")
example.coords <- system.file("examples/example_coordinates.csv",
                            package="rangeExpansion")

raw.data <- load.data.snapp("examples/example_snp.snapp", 
                            "examples/example_coordinates.csv", 
                            sep=',', ploidy=ploidy)                                
region <- list(NULL, c("REGION_1", "REGION_2"))

pop <- make.pop(raw.data, ploidy)
psi <- get.all.psi(pop)

res <- run.regions(region=region, pop=pop, psi=psi, xlen=10,ylen=20)

summary(res)
plot(res)
}

