% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/re_functions2.r
\name{set.outgroups}
\alias{set.outgroups}
\title{Sets columns of outgroup individuals}
\usage{
set.outgroups(data, ploidy = 2)
}
\arguments{
\item{data}{output from check missing}

\item{ploidy}{the number of copies per individual}
}
\value{
An object with entries 'genotypes' and 'coords' containing the
   input data sets
}
\description{
As psi requires the derived state of an allele to be known, this function
provides an easy way to polarize SNP with outgroup individuals. From the
outgroup columns in the coord file, outgroup individuals are removed
and SNPs swithced if necessary
}
\examples{


example.data <- system.file("examples/example_snp.snapp",
                            package="rangeExpansion")
example.coords <- system.file("examples/example_coordinates.csv",
                            package="rangeExpansion")

region <- list(NULL, c("REGION_1", "REGION_2"))
ploidy <- 2
raw.data <- load.data.snapp(example.data,
                            example.coords,
                            sep=',', ploidy=ploidy)                                

pop <- make.pop(raw.data, ploidy)
psi <- get.all.psi(pop)

res <- run.regions(region=region, pop=pop, psi=psi, xlen=10,ylen=20)

summary(res)
plot(res)
}

